split logic to new file
